// src/types/invoice.ts
export type InvoiceStatus = "draft" | "pending" | "paid";

export interface InvoiceItem {
  id: string;
  name: string;
  quantity: number;
  price: number;
  total: number;
}

export interface Invoice {
  id: string;
  createdAt: string; // ISO invoice date
  paymentDue: string; // ISO due date
  paymentTerms: number; // days
  description: string;
  status: InvoiceStatus;
  clientName: string;
  clientEmail: string;
  senderAddress: {
    street: string;
    city: string;
    postCode: string;
    country: string;
  };
  clientAddress: {
    street: string;
    city: string;
    postCode: string;
    country: string;
  };
  items: InvoiceItem[];
  total: number;
}